Notification restore revamp to use Worker now * **Deleted Files:**   * `NotificationRestorer`   * `NotificationRestoreService`   * `RestoreJobService`   * `RestoreKickoffJobService`   * Services all removed from `AndroidManifest` as well since `Worker` acts as a replacement * `UpgradeReceiver` and `BootUpReceiver` now kickoff notification restoring through `OSNotificationRestoreWorkManager`
* Now will plan on rebasing this PR behind `feature/add_worker_class_for_processing_and_foreground_handler`
  * This is so we can restore notifications using the new `OSNotificationWorkManager` class
