+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/OneSignal.java
@@ -800,7 +800,7 @@ public class OneSignal {
       inForeground = isContextActivity(context);
       if (inForeground) {
          ActivityLifecycleHandler.curActivity = (Activity) context;
-         NotificationRestorer.asyncRestore(appContext);
+         OSNotificationRestoreWorkManager.beginEnqueueingWork(context);
          FocusTimeController.getInstance().appForegrounded();
       }
       else
@@ -1178,7 +1178,7 @@ public class OneSignal {
       if (trackGooglePurchase != null)
          trackGooglePurchase.trackIAP();
 
-      NotificationRestorer.asyncRestore(appContext);
+      OSNotificationRestoreWorkManager.beginEnqueueingWork(appContext);
 
       getCurrentPermissionState(appContext).refreshAsTo();
