+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/UpgradeReceiver.java
@@ -37,10 +37,13 @@ public class UpgradeReceiver extends BroadcastReceiver {
    @Override
    public void onReceive(Context context, Intent intent) {
 
+      // TODO: Now that we arent restoring like we use to, think we can remove this? Ill do some
+      //  testing and look at the issue but maybe someone has a answer or rems what directly
+      //  was causing this issue
       // Return early if using Android 7.0 due to upgrade restore crash (#263)
       if (Build.VERSION.SDK_INT == Build.VERSION_CODES.N)
          return;
-      
-      NotificationRestorer.startDelayedRestoreTaskFromReceiver(context);
+
+      OSNotificationRestoreWorkManager.beginEnqueueingWork(context);
    }
 }
